requirements.txt
flask==3.1.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
plotly==5.24.1
requests==2.32.3
gunicorn==23.0.0
